src/requirements.txt
bip-utils>=2.9.0
aiogram>=3.21.0
tronpy
fastapi
uvicorn[standard]
sqlalchemy
qrcode
python-dotenv
pillow
PyJWT>=2.8.0

#tests
pytest
pytest-asyncio